do not enforce string count for OemStrings
Although there is an explicit count for the number of OemStrings,
firmware manufacturers do not necessarily keep it up to date with the
actual number of strings. Any code that is using the OemStrings
structure (with the strings iterator) will be able to walk all of the
strings that are present if we do not enforce checking the count. Fixes