ignore doc-test on non-exported macro
Rust 1.62 made a breaking change and started running doc-tests on
non-exported macros (of which the layout macro is) which caused the
doc-test to fail (because it cannot use the macro since it isn't
exported and there is no way to run a doc-test on that). Putting an
ignore on the code block prevents it from being interpreted as a
doc-test.